+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 - Fix undo/redo for new item not reverting to empty body after save
 - Fix undo/redo not working properly with complex fields (dateline, category, subject)
 - Fix images uploaded into private space are not visible
+- Fix desk selection is monitoring if not default desk
+- Fix elastic highlight config to use `features` key.
+- Fix metadataview for slugline to display elastic highlight
+- Fix elastic highlight for saved searches.
 
 ## [1.0.0-beta1] 2016-04-26